requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.0.3
django-allauth==0.54.0
django-cors-headers==4.0.0
django-dotenv==1.4.2
django-filter==23.2
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
Markdown==3.4.3
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.1
urllib3==2.0.2
whitenoise==6.4.0
zipp==3.15.0
